hardware/cabinet-client/pyproject.toml
[project]
name = "cabinet-client"
version = "1.0.0"
description = "Cabinet program for the cabinet computer"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "numpy>=2.4.3",
    "obs-websocket-py>=1.0",
    "opencv-python>=4.13.0.92",
    "pyinstaller>=6.19.0",
    "pyserial>=3.5",
    "requests>=2.32.5",
    "websockets>=16.0",
]
